+++ lib/gandi_v5/domain/transfer_in.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class GandiV5
+  class Domain
+    # Manage a domain's transfer to Gandi.
+    class TransferIn
+    end
+  end
+end