+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gandi V5 API Gem Changelog
 
+* Add transferring a domain to Gandi (I don't have any domains outside Gandi to test this myself so it's possible I've misread the docs and a bug is waiting to be found, please add an issue if I have):
+  * GandiV5::Domain::TransferIn::Availabillity:
+    * .fetch(fqdn, auth_code = nil)
+
 ## Version 0.8.0
 
 * Domain: